Initial Commit
Commit with all the initial code. Contains menu, and a simple level,
with two characters from the player side and one character from the
computer side. The computer automatically generates that one character.
Currently, the characters do not detect the towers. Also, the costs of
the characters are not working (bug)